separate cli into submodule
