+++ pyproject.toml
@@ -14,13 +14,13 @@ license = "MIT"
 repository = "https://github.com/SuperLesson/SuperLesson"
 
 [tool.poetry.scripts]
-superlesson = "superlesson:main"
-transcribe = "superlesson:transcribe_step"
-tmarks = "superlesson:tmarks_step"
-verify = "superlesson:verify_step"
-replace = "superlesson:replace_step"
-improve = "superlesson:improve_step"
-annotate = "superlesson:annotate_step"
+superlesson = "superlesson.cli:main"
+transcribe = "superlesson.cli:transcribe_step"
+tmarks = "superlesson.cli:tmarks_step"
+verify = "superlesson.cli:verify_step"
+replace = "superlesson.cli:replace_step"
+improve = "superlesson.cli:improve_step"
+annotate = "superlesson.cli:annotate_step"
 
 [tool.poetry.dependencies]
 python = "^3.9"
