+++ superlesson/__main__.py
@@ -1,4 +1,4 @@
-from . import main
+from .cli import main
 
 
 if __name__ == "__main__":
